requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.27
opencv-python==4.7.0.72
numpy==1.24.4
Werkzeug==2.0.3
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
scikit-learn
Pillow>=7.1.2
gdown